requirements.txt
datasets
transformers
torch
trl
peft
bitsandbytes
accelerate
colorama
python-dotenv
